# [torch.compile] use size tuning for specific sizes