@youkaichao 2024-12-05
I don't have exact measurement. it seems the first compilation takes less than 1 minute, and later the auto-tune results are cached somewhere and cost nothing.